Added examples of Venue classes + trivial style edits
Followed the Awards example to create Venue example so that there is no ambiguity.  Followed format precisely - including use of "venue" and "venues" following the different use of "award" and "awards".  This change made assuming the earlier instruction "syntax follows the awards example" is literal.  If this is correct, I can repeat addition for funders, institutions, and organizations so that all are explicit.